pj: merge pj_poi.py and legacy_pj_poi.py
