+++ idunn/places/__init__.py
@@ -3,8 +3,7 @@ from .admin import Admin
 from .poi import POI, BragiPOI
 from .street import Street
 from .place import Place
-from .pj_poi import PjPOI
-from .legacy_pj_poi import LegacyPjPOI
+from .pj_poi import PjPOI, LegacyPjPOI
 from .latlon import Latlon
 from .event import Event
