+++ idunn/places/models/pj_find.py
@@ -1,6 +1,6 @@
-# generated by datamodel-codegen:
-#   filename:  spec(1).yaml
-#   timestamp: 2020-11-26T07:58:40+00:00
+# based on model generated by datamodel-codegen:
+#   filename:  spec.yaml
+#   timestamp: 2020-11-26t07:58:40+00:00
 
 from __future__ import annotations
